Port #3/2 from unixlike-builds branch: made construtors capable of accepting a single argument explicit. Port #4/3 from unixlike-builds branch: fixed some warnings treated as errors (-Wconversion -Wsign-conversion). Port #5/2 from unixlike-builds branch: fixed some typos, and corrected grammar. Port #7/2 from unixlike-builds branch: fixed tests. - fixed tests relying on order of files when using the recursive directory iterator. Port #8/2 from unixlike-builds branch: mainly suggestions from CLion (clang-tidy). - add [[nodiscard]]. - add const to methods. - add override keyword to derived classes virtual methods. - change some members and methods to static. - remove virtual keyword from derived classes virtual methods.

+++ include/the_modern_c++_challenge/chapter_02_language_features/ipv4.h
@@ -2,7 +2,7 @@
 
 #include <algorithm>  // for_each
 #include <array>
-#include <charconv>
+#include <charconv>  // from_chars
 #include <cstdint>  // uint8_t, uint32_t
 #include <fmt/format.h>
 #include <fmt/ostream.h>
@@ -35,7 +35,7 @@ namespace tmcppc::network {
     // std::array is movable if its elements are movable
     // That makes implementation of move constructor and move assignment operator unnecessary for the IPv4 class,
     // since all the members are movable
-    // Copy construction, copy assignment operator and destructor are also unnecesary
+    // Copy construction, copy assignment operator and destructor are also unnecessary
     //
     // operator<< and operator>> implemented as non-member functions and friends
     class ipv4 {
@@ -53,7 +53,12 @@ namespace tmcppc::network {
         {}
 
         explicit constexpr ipv4(std::uint32_t address)
-            : octets_{ (address >> 24) & 0xff, (address >> 16) & 0xff, (address >> 8) & 0xff, address & 0xff }
+            : octets_{
+                static_cast<uint8_t>((address >> 24) & 0xff),
+                static_cast<uint8_t>((address >> 16) & 0xff),
+                static_cast<uint8_t>((address >> 8) & 0xff),
+                static_cast<uint8_t>(address & 0xff)
+            }
         {}
 
         constexpr ipv4(std::uint8_t o0, std::uint8_t o1, std::uint8_t o2, std::uint8_t o3)
@@ -67,7 +72,7 @@ namespace tmcppc::network {
         }
 
         ipv4& operator++() {
-            *this = ipv4{ this->to_ulong() + 1 };
+            *this = ipv4{ static_cast<uint32_t>(this->to_ulong() + 1) };
             return *this;
         }
